Move list and print func to resource pkg
